fix(yarn): pin scrypt to latest github version
A [PR was merged in](https://github.com/barrysteyn/node-scrypt/pull/197) last year to add support for Node 12+.  The maintainer never published a new version on NPM though, so this fix isn't doing much.

This commit instead pins the version of scrypt to the latest git revision, which includes the fix.  This might mean that we miss out on future updates, but the module is [already deprecated](https://github.com/barrysteyn/node-scrypt#scrypt-for-node) so the risk of that is low.